Added input manager
