+++ Classes/Lesson.cs
@@ -6,6 +6,7 @@ using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using ThirdTeam_Study.ListTypes;
 using ThirdTeam_Study.Managers;
 using static System.Formats.Asn1.AsnWriter;
